+++ code.verse/src/main/java/app/code/verse/application/usecases/UpdateEmployee.java
@@ -1,4 +1,4 @@
-package app.code.verse.domain.services;
+package app.code.verse.application.usecases;
 
 import app.code.verse.domain.model.Employee;
 import app.code.verse.domain.model.utils.ValidateDataUtil;
