+++ code.verse/src/main/java/app/code/verse/domain/services/RegisterPatient.java
@@ -1,6 +1,6 @@
 package app.code.verse.domain.services;
 
-import app.code.verse.domain.model.Employee;
+import app.code.verse.application.usecases.FindPatient;
 import app.code.verse.domain.model.Patient;
 import app.code.verse.domain.model.utils.ValidateDataUtil;
 import app.code.verse.domain.ports.PatientPort;
